requirements.txt
# torch >= 2.0.0
tqdm
rich
ninja
numpy
pandas
scipy
scikit-learn
matplotlib
imageio
imageio-ffmpeg
omegaconf

einops
plyfile
pygltflib

# for stable-diffusion
huggingface_hub
accelerate
transformers

# for dmtet and mesh export
xatlas
trimesh
PyMCubes==0.1.4
# pymeshlab==0.2
pymeshlab
kornia
opencv-python
torchmetrics
